requirements.txt
astor==0.8.1
asttokens==2.4.0
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
fonttools==4.43.1
formulaic==0.6.6
future==0.18.3
idna==3.4
interface-meta==1.3.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lifelines==0.27.8
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
ngboost==0.3.12
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
wrapt==1.15.0
